+++ src/webapp/src/api/sql.js
@@ -6,9 +6,9 @@ export default {
     return axios.get(utils.apiUrl('sql/get', 'filter-options'))
   },
 
-  getSql(namespace, model, design, pipeline, data) {
+  getSql(namespace, model, design, data) {
     return axios.post(
-      utils.apiUrl('sql/get', `${namespace}/${model}/${design}/${pipeline}`),
+      utils.apiUrl('sql/get', `${namespace}/${model}/${design}`),
       data
     )
   }